pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wolkenwalze_plugin_sdk"
# Do not change, the version is automatically updated in CI.
version = "0.0.0"
authors = [
]
description = "Plugin SDK for the Wolkenwalze workflow engine"
readme = "README.md"
license = { text="Apache-2.0" }
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
  'typing_extensions ~= 4.3',
  'PyYAML ~= 6.0',
]

[project.urls]
"Homepage" = "https://github.com/wolkenwalze/wolkenwalze-python-sdk"
"Bug Tracker" = "https://github.com/wolkenwalze/wolkenwalze-python-sdk/issues"
